pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-drone-env"
version = "0.1.0"
description = "Agentic Drone PID Tuner Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "numpy>=1.21.0",
    "transformers>=4.30.0",
    "trl>=0.8.6"
]

[tool.setuptools]
packages = ["drone_env"]

[project.scripts]
drone-env-server = "drone_env.server.app:app"

[project.entry-points."openenv.envs"]
drone_env = "drone_env.server.app:app"
